Refactor kube controller
