+++ pilot/pkg/serviceregistry/kube/controller/endpoints.go
@@ -162,10 +162,6 @@ func (e *endpointsController) getInformer() cache.SharedIndexInformer {
 }
 
 func (e *endpointsController) onEvent(curr interface{}, event model.Event) error {
-	if err := e.c.checkReadyForEvents(); err != nil {
-		return err
-	}
-
 	ep, ok := curr.(*v1.Endpoints)
 	if !ok {
 		tombstone, ok := curr.(cache.DeletedFinalStateUnknown)
