+++ pilot/pkg/serviceregistry/kube/controller/util.go
@@ -15,8 +15,20 @@
 package controller
 
 import (
+	"encoding/json"
+	"fmt"
+
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	klabels "k8s.io/apimachinery/pkg/labels"
+	"k8s.io/apimachinery/pkg/util/intstr"
+	listerv1 "k8s.io/client-go/listers/core/v1"
+
+	"istio.io/pkg/log"
+
+	"istio.io/istio/pilot/pkg/model"
+	"istio.io/istio/pilot/pkg/serviceregistry/kube"
+	"istio.io/istio/pkg/config/labels"
 )
 
 func hasProxyIP(addresses []v1.EndpointAddress, proxyIP string) bool {
@@ -36,4 +48,131 @@ func getLabelValue(metadata metav1.Object, label string, fallBackLabel string) s
 	}
 
 	return labels[fallBackLabel]
+}
+
+func createUID(podName, namespace string) string {
+	return "kubernetes://" + podName + "." + namespace
+}
+
+// Forked from Kubernetes k8s.io/kubernetes/pkg/api/v1/pod
+// FindPort locates the container port for the given pod and portName.  If the
+// targetPort is a number, use that.  If the targetPort is a string, look that
+// string up in all named ports in all containers in the target pod.  If no
+// match is found, fail.
+func FindPort(pod *v1.Pod, svcPort *v1.ServicePort) (int, error) {
+	portName := svcPort.TargetPort
+	switch portName.Type {
+	case intstr.String:
+		name := portName.StrVal
+		for _, container := range pod.Spec.Containers {
+			for _, port := range container.Ports {
+				if port.Name == name && port.Protocol == svcPort.Protocol {
+					return int(port.ContainerPort), nil
+				}
+			}
+		}
+	case intstr.Int:
+		return portName.IntValue(), nil
+	}
+
+	return 0, fmt.Errorf("no suitable port for manifest: %s", pod.UID)
+}
+
+// findPortFromMetadata resolves the TargetPort of a Service Port, by reading the Pod spec.
+func findPortFromMetadata(svcPort v1.ServicePort, podPorts []model.PodPort) (int, error) {
+	target := svcPort.TargetPort
+
+	switch target.Type {
+	case intstr.String:
+		name := target.StrVal
+		for _, port := range podPorts {
+			if port.Name == name && port.Protocol == string(svcPort.Protocol) {
+				return port.ContainerPort, nil
+			}
+		}
+	case intstr.Int:
+		// For a direct reference we can just return the port number
+		return target.IntValue(), nil
+	}
+
+	return 0, fmt.Errorf("no matching port found for %+v", svcPort)
+}
+
+func getPodServices(s listerv1.ServiceLister, pod *v1.Pod) ([]*v1.Service, error) {
+	allServices, err := s.Services(pod.Namespace).List(klabels.Everything())
+	if err != nil {
+		return nil, err
+	}
+
+	var services []*v1.Service
+	for _, service := range allServices {
+		if service.Spec.Selector == nil {
+			// services with nil selectors match nothing, not everything.
+			continue
+		}
+		selector := klabels.Set(service.Spec.Selector).AsSelectorPreValidated()
+		if selector.Matches(klabels.Set(pod.Labels)) {
+			services = append(services, service)
+		}
+	}
+
+	return services, nil
+}
+
+func portsEqual(a, b []v1.EndpointPort) bool {
+	if len(a) != len(b) {
+		return false
+	}
+
+	for i := range a {
+		if a[i].Name != b[i].Name || a[i].Port != b[i].Port || a[i].Protocol != b[i].Protocol ||
+			ptrValueOrEmpty(a[i].AppProtocol) != ptrValueOrEmpty(b[i].AppProtocol) {
+			return false
+		}
+	}
+
+	return true
+}
+
+func addressesEqual(a, b []v1.EndpointAddress) bool {
+	if len(a) != len(b) {
+		return false
+	}
+
+	for i := range a {
+		if a[i].IP != b[i].IP || a[i].Hostname != b[i].Hostname ||
+			ptrValueOrEmpty(a[i].NodeName) != ptrValueOrEmpty(b[i].NodeName) {
+			return false
+		}
+	}
+
+	return true
+}
+
+func ptrValueOrEmpty(ptr *string) string {
+	if ptr != nil {
+		return *ptr
+	}
+	return ""
+}
+
+func getNodeSelectorsForService(svc *v1.Service) labels.Instance {
+	if nodeSelector := svc.Annotations[kube.NodeSelectorAnnotation]; nodeSelector != "" {
+		var nodeSelectorKV map[string]string
+		if err := json.Unmarshal([]byte(nodeSelector), &nodeSelectorKV); err != nil {
+			log.Debugf("failed to unmarshal node selector annotation value for service %s.%s: %v",
+				svc.Name, svc.Namespace, err)
+		}
+		return nodeSelectorKV
+	}
+	return nil
+}
+
+func nodeEquals(a, b kubernetesNode) bool {
+	return a.address == b.address && a.labels.Equals(b.labels)
+}
+
+func isNodePortGatewayService(svc *v1.Service) bool {
+	_, ok := svc.Annotations[kube.NodeSelectorAnnotation]
+	return ok && svc.Spec.Type == v1.ServiceTypeNodePort
 }
